Refactor and enhance integration logic: change `num_steps` to `size_t` to prevent overflow, improve precision in `depthAveragedIrradiance`, update naming in CMakeLists, clean up formatting across tests, and document RK2/RK4 refactor design in `AGENTS.md`.

+++ cli/main.cpp
@@ -12,7 +12,7 @@ int main() {
     constexpr double k = 0.2;
     constexpr double Kd = 0.3; //
     constexpr double dt = 0.01;
-    constexpr int num_steps = 1000;
+    constexpr size_t num_steps = 1000;
     const auto geometry = ReactorGeometry(depth, I0, k);
